build: the warning level is /W4, and the tree is quiet at it
CMake seeds CMAKE_CXX_FLAGS with /W3 and the root CMakeLists named no level
of its own, so /W3 is what the tree had been clean at. /W4 found 613
warnings across 589 sites, every one of them first party - the SYSTEM
includes for boost, SDL and Vulkan keep vcpkg's 80,000 headers quiet
already, so no external filter was needed.

Two files are exempt rather than fixed. moya's RenderMan.cxx declares the
whole RenderMan Interface and leaves most entry points empty, so its
unreferenced parameters were 501 of the 613 - the target takes /wd4100,
since naming them out would only be undone as each function grows a body.
voxel's noiseutils.cpp is Jason Bevins' libnoise utility layer vendored
verbatim and takes /wd4100, /wd4189 and /wd4244 for the reason cpplint
already skips it. That one is set twice: a source file property is scoped
to the directory that sets it, and voxel/tests compiles the same file.

Four of the rest were defects rather than noise.

The jpeg reader leaked on any decode failure. img was constructed after the
setjmp point, so libjpeg's longjmp out of jpeg_read_scanlines skipped its
destructor; the shared_ptr is now declared above the setjmp and unwinds on
the error return. That is what C4611 was pointing at, and the pragma that
silences it records the invariant the declaration now satisfies.

BitmapFont::CharDescriptor was a POD with no initialisers. A char line in a
.fnt names its fields by key, so any field an abbreviated line omitted was
copied into the charset indeterminate - and channel_ was never parsed at
all, the chnl branch being empty, so it was always garbage.

Project::load narrowed uint64_t to brep::Index at four references per edge
and one per face, so a document naming an index above 2^32 wrapped into a
valid looking one. index() now yields brep::Index and rejects what does not
fit, which is the same rule the loader already applies to a double or a
negative. refers() follows the type.

The tga writer put a uint32_t width and height into the 16 bit fields the
format has. An image the format cannot describe is refused now.

Two more were wrong without being reachable: wcsdupstr allocated its length
in bytes and its array in elements, over-allocating twofold, and the png
writer read color_type uninitialised if Format were ever neither RGB nor
RGBA.

Image::bpp() returned unsigned int while bpp_ is uint8_t, so five callers
narrowed on assignment. The getter reports the type it stores, which is
what the constructor already took. TextBuffer::addIndex takes unsigned int
for the same reason - the index buffer holds what the draw takes.

The remainder is style: unused parameters lose their names to /* name */ at
19 sites, dead locals and one unreachable return go, strerror and wcscpy
become the _s forms, and the shadowed ball_dir in PongScene scopes into the
AI block. The tga signature goes to memcpy rather than strncpy_s - the
field is exactly 16 bytes with no terminator, and _TRUNCATE would have
written 15 and a NUL.

A clean build of all 403 targets reports no warnings. The suite passes 20
of 20 and cpplint reports only the indent_namespace noise.

+++ api/font/TextBuffer.cxx
@@ -16,7 +16,7 @@ namespace v3d::font {
         vertices_.push_back(vertex);
     }
 
-    void TextBuffer::addIndex(size_t index) {
+    void TextBuffer::addIndex(unsigned int index) {
         indices_.push_back(index);
     }
